JS to display Tags
I'm just guessing regarding "item.Tags" -- wasn't sure where to look to verify that (so this is untested!). For now I've just gone with using the same approach as how Genres are handled of using bullets to separate them (so I mostly just copied and modified the Genres code), but I also think it would be fine to use the gray rounded-corner button look (like the tags appear in the server web interface, and like the media info details appear in ET). I might lean a little more towards the gray button look personally, but this was easier/quicker to implement. :)